+++ src/compiler/cpp_generator.cc
@@ -143,7 +143,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/support/client_callback.h",
         "grpcpp/client_context.h",
         "grpcpp/impl/codegen/completion_queue.h",
-        "grpcpp/impl/codegen/message_allocator.h",
+        "grpcpp/support/message_allocator.h",
         "grpcpp/impl/codegen/method_handler.h",
         "grpcpp/impl/codegen/proto_utils.h",
         "grpcpp/impl/codegen/rpc_method.h",
@@ -1653,7 +1653,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "grpcpp/impl/codegen/channel_interface.h",
         "grpcpp/impl/codegen/client_unary_call.h",
         "grpcpp/support/client_callback.h",
-        "grpcpp/impl/codegen/message_allocator.h",
+        "grpcpp/support/message_allocator.h",
         "grpcpp/impl/codegen/method_handler.h",
         "grpcpp/impl/codegen/rpc_service_method.h",
         "grpcpp/impl/codegen/server_callback.h",
